task_3 fixed remarks: added index.ts in the components folder and simplified imports; removed comment; moved on import styles to last line
