+++ src/components/MovieCard/MovieCard.module.scss
@@ -55,7 +55,6 @@
     position: relative;
     bottom: 500px;
     left: 273px;
-    //display: flex;
     justify-content: center;
     align-items: center;
     display: none;
