+++ src/components/index.ts
@@ -1,0 +1,13 @@
+export * from './ErrorBoundary';
+export * from './Main';
+export * from './Button';
+export * from './Divider';
+export * from './Footer';
+export * from './MovieList';
+export * from './MovieCard';
+export * from './SearchForm';
+export * from './Header';
+export * from './CountMovie';
+export * from './MovieFilter';
+export * from './MovieSort';
+export * from './NetflixRoulette';
